requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
numpy>=1.24.0,<2.0.0
langchain-google-genai
langchain-community
langchain-chroma
pysqlite3-binary
pydantic
pypdf
fastapi
uvicorn
mangum
pysqlite3-binary